refactor js package functions
