--- packages/js/src/index.ts
@@ -1,2 +1,0 @@
-export { default as formbricksApp } from "./app";
-export { default as formbricksWebsite } from "./website";
